ref(rotues): Extract custom Route function to route component
